+++ extensions/rxjava3-extensions/src/test/kotlin/com/squareup/sqldelight/runtime/rx3/QueryObservableTest.kt
@@ -1,9 +1,9 @@
-package com.squareup.sqldelight.runtime.rx
+package com.squareup.sqldelight.runtime.rx3
 
 import com.squareup.sqldelight.Query
 import com.squareup.sqldelight.internal.copyOnWriteList
-import com.squareup.sqldelight.runtime.rx.Employee.Companion.SELECT_EMPLOYEES
-import com.squareup.sqldelight.runtime.rx.TestDb.Companion.TABLE_EMPLOYEE
+import com.squareup.sqldelight.runtime.rx3.Employee.Companion.SELECT_EMPLOYEES
+import com.squareup.sqldelight.runtime.rx3.TestDb.Companion.TABLE_EMPLOYEE
 import io.reactivex.rxjava3.schedulers.Schedulers
 import org.junit.Test
