+++ extensions/rxjava3-extensions/src/test/kotlin/com/squareup/sqldelight/runtime/rx3/RecordingObserver.kt
@@ -13,7 +13,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package com.squareup.sqldelight.runtime.rx
+package com.squareup.sqldelight.runtime.rx3
 
 import com.google.common.truth.Truth.assertThat
 import com.squareup.sqldelight.Query
