+++ extensions/rxjava3-extensions/src/test/kotlin/com/squareup/sqldelight/runtime/rx3/TestDb.kt
@@ -1,12 +1,12 @@
-package com.squareup.sqldelight.runtime.rx
+package com.squareup.sqldelight.runtime.rx3
 
 import com.squareup.sqldelight.Query
 import com.squareup.sqldelight.TransacterImpl
 import com.squareup.sqldelight.db.SqlCursor
 import com.squareup.sqldelight.db.SqlDriver
 import com.squareup.sqldelight.internal.copyOnWriteList
-import com.squareup.sqldelight.runtime.rx.TestDb.Companion.TABLE_EMPLOYEE
-import com.squareup.sqldelight.runtime.rx.TestDb.Companion.TABLE_MANAGER
+import com.squareup.sqldelight.runtime.rx3.TestDb.Companion.TABLE_EMPLOYEE
+import com.squareup.sqldelight.runtime.rx3.TestDb.Companion.TABLE_MANAGER
 import com.squareup.sqldelight.sqlite.driver.JdbcSqliteDriver
 import com.squareup.sqldelight.sqlite.driver.JdbcSqliteDriver.Companion.IN_MEMORY
